Rename identity-provider.js to identity-provider-interface.js
